fix: 获取Motd时人数过高可能异常 fix: 删除不必要的判断

+++ Serein/Universal/Base/Motd/Motdje.cs
@@ -107,8 +107,8 @@ namespace Serein.Base.Motd
             if (!string.IsNullOrEmpty(Origin))
             {
                 JObject jsonObject = (JObject)JsonConvert.DeserializeObject(Origin);
-                OnlinePlayer = int.Parse(jsonObject["players"]["online"].ToString());
-                MaxPlayer = int.Parse(jsonObject["players"]["max"].ToString());
+                OnlinePlayer = long.Parse(jsonObject["players"]["online"].ToString());
+                MaxPlayer = long.Parse(jsonObject["players"]["max"].ToString());
                 Version = jsonObject["version"]["name"].ToString();
                 Protocol = jsonObject["version"]["protocol"].ToString();
                 if (jsonObject["description"]["text"] != null)
